Fix inconsitent named tests
